+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Create `form` and `view` display configurations for new bundles
+- Add further options for node bundles
 
 ## [0.1.2] - 2019-07-25
 ### Changed