+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0]
 ### Added
 - Create `form` and `view` display configurations for new bundles
 - Add further options for node bundles
@@ -28,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add site_builder_console:field:create command
 - Add site_builder_console:field:delete command
 
-[Unreleased]: https://github.com/wongjn/site_builder_console/compare/0.1.2...HEAD
+[Unreleased]: https://github.com/wongjn/site_builder_console/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/wongjn/site_builder_console/compare/0.1.2...0.2.0
 [0.1.2]: https://github.com/wongjn/site_builder_console/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/wongjn/site_builder_console/compare/0.1.0...0.1.1